requirements.txt
pandas
networkx
python-levenshtein
lxml.html
lxml.etree
requests
